ISSUE-1096 Move TWP settings consumer from `jmap-extensions-rabbitmq` to `saas/saas-rabbitmq` module
So the SaaS connector can reuse the same RabbitMQ of TWP settings consumer.
