+++ tmail-backend/guice/distributed/pom.xml
@@ -62,6 +62,10 @@
             <scope>test</scope>
         </dependency>
         <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>tmail-saas-rabbitmq</artifactId>
+        </dependency>
+        <dependency>
             <groupId>${james.groupId}</groupId>
             <artifactId>apache-james-backends-redis</artifactId>
         </dependency>
