+++ tmail-backend/jmap/extensions-api/src/main/scala/com/linagora/tmail/james/jmap/settings/JmapSettingsUtil.scala
